Move IPAddress to application
